% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancemeasures.R
\name{distances}
\alias{distances}
\title{distances}
\usage{
distances(data, model, plots = TRUE)
}
\arguments{
\item{data}{A Dataframe}

\item{model}{A linear model of class lm}

\item{plots}{A boolean parameter to supress plots}
}
\value{
a list with names entries (cooks,dffits,hadi)
}
\description{
distances
}
