% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distancemeasures.R
\name{distances}
\alias{distances}
\title{distances}
\usage{
distances(data, model, plots = TRUE)
}
\arguments{
\item{data}{A Dataframe}

\item{model}{A linear model of class lm}

\item{plots}{A boolean parameter to supress plots}
}
\value{
a list with names entries (cooks,dffits,hadi)
}
\description{
distances
}
\examples{
{ x1 <- rnorm(100, 50, 9);
y1 <-5*x1+1;
randomdata <- data.frame(y1, x1);
model <- lm(y1 ~ x1, data = randomdata);
distances(randomdata, model, plots = TRUE)
}
}
